+++ src/lib/auth/authSlice.js
@@ -1,6 +1,26 @@
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 import { API_ENDPOINT } from '../../config';
 
+// Async thunk for Twitter login
+export const twitterLogin = createAsyncThunk(
+  'auth/twitterLogin',
+  async (_, { rejectWithValue }) => {
+    try {
+      const response = await fetch(`${API_ENDPOINT}/auth/twitter/login`);
+      const data = await response.json();
+
+      if (data.authorization_url) {
+        return data.authorization_url;
+        // window.location.href = data.authorization_url;
+      } else {
+        return rejectWithValue(`Error logging in with Twitter ${data}`);
+      }
+    } catch (error) {
+      return rejectWithValue(`Error logging in with Twitter: ${error}`);
+    }
+  }
+);
+
 // Async thunk for fetching user data
 export const fetchUserData = createAsyncThunk(
   'user/fetchUserData',
@@ -26,12 +46,14 @@ export const fetchUserData = createAsyncThunk(
   }
 );
 
-const userSlice = createSlice({
+const authSlice = createSlice({
   name: 'auth',
   initialState: {
     user: JSON.parse(localStorage.getItem('user')) || null,
     token: localStorage.getItem('token') || null,
+    authorizationUrl: null,
     isLoading: false,
+    isMenuOpen: false,
     error: null,
   },
   reducers: {
@@ -51,6 +73,9 @@ const userSlice = createSlice({
         localStorage.removeItem('user');
       }
     },
+    setIsMenuOpen: (state, action) => {
+      state.isMenuOpen = action.payload;
+    },
   },
   extraReducers: (builder) => {
     builder
@@ -68,10 +93,17 @@ const userSlice = createSlice({
         state.error = action.payload;
         state.user = null;
         localStorage.removeItem('user');
+      })
+      .addCase(twitterLogin.rejected, (state, action) => {
+        console.error(action.payload);
+      })
+      .addCase(twitterLogin.fulfilled, (state, action) => {
+        state.authorizationUrl = action.payload;
       });
   },
 });
 
-export const { setToken, clearToken, updateUser } = userSlice.actions;
+export const { setToken, clearToken, updateUser, setIsMenuOpen } =
+  authSlice.actions;
 
-export default userSlice.reducer;
+export default authSlice.reducer;